+++ src/core/datastore/persisters/browser/index.js
@@ -1,0 +1,2 @@
+export * from './websql-key-value-store-persister';
+export * from './indexeddb-key-value-store-persister';
