+++ src/core/datastore/persisters/browser/websql-key-value-store-persister.js
@@ -1,6 +1,6 @@
-import { KinveyError } from '../../errors';
+import { KinveyError } from '../../../errors';
 
-import { KeyValueStorePersister } from './key-value-store-persister';
+import { KeyValueStorePersister } from './../key-value-store-persister';
 import { webSqlCollectionsMaster, webSqlDatabaseSize } from '../utils';
 import { ensureArray } from '../../utils';
