+++ src/core/datastore/persisters/index.js
@@ -3,5 +3,4 @@ export * from './key-value-persister';
 export * from './persister-provider';
 export * from './memory-key-value-persister';
 export * from './key-value-store-persister';
-export * from './websql-key-value-store-persister';
-export * from './indexeddb-key-value-store-persister';
+export * from './browser';
